chore: repair PR #195 with architectural improvements and fixes (#313)
