+++ src/components/ui/ContentCard.tsx
@@ -5,7 +5,7 @@ import { readingTime } from '@/lib/content';
 import { CardImagePlaceholder } from '@/components/ui/CardImagePlaceholder';
 import { Skeleton } from '@/components/ui/Skeleton';
 import { CategoryPlaceholder } from '@/components/ui/CategoryPlaceholder';
-import { useImage } from '@/hooks/useImage';
+
 
 interface ContentCardProps extends Partial<HTMLMotionProps<"a">> {
   slug: string;
@@ -40,7 +40,6 @@ export function ContentCardSkeleton() {
 
 export function ContentCard({ slug, title, category, excerpt, date, image, basePath, content, ...motionProps }: ContentCardProps) {
   const rt = readingTime(content, excerpt);
-  const { imgError, handleImgError } = useImage();
 
   return (
     <Box 
@@ -55,7 +54,7 @@ export function ContentCard({ slug, title, category, excerpt, date, image, baseP
             src={image} 
             alt={title} 
             className="w-full h-full object-cover group-hover:scale-105 transition-transform duration-700"
-            onError={handleImgError}
+            onError={(e) => { e.currentTarget.onerror = null; e.currentTarget.style.display = 'none'; }}
           />
         ) : (
           <Box className="w-full h-full flex flex-col">
