+++ src/features/contact/components/FormField.tsx
@@ -1,10 +1,10 @@
-import React, { useId } from 'react';
+import { useId, cloneElement, ReactElement } from 'react';
 import { Box, Stack, Text } from '@/layouts/Primitives';
 
 interface FormFieldProps {
   label: string;
   error?: string;
-  children: React.ReactElement;
+  children: ReactElement;
 }
 
 export function FormField({ label, error, children }: FormFieldProps) {
@@ -23,7 +23,7 @@ export function FormField({ label, error, children }: FormFieldProps) {
           </Text>
         )}
       </Box>
-      {React.cloneElement(children, {
+      {cloneElement(children, {
         id,
         'aria-describedby': error ? errorId : undefined,
         'aria-invalid': !!error
